pyproject.toml
[tool.poetry]
name = "app"
version = "0.1.0"
description = ""
authors = ["Nathan Johnson <nathanjohnson320@gmail.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.111.1"
pydantic-settings = "^2.3.4"
sqlalchemy = "^2.0.31"
fastapi-pagination = "^0.12.26"
psycopg2-binary = "^2.9.9"
alembic = "^1.13.2"
argon2-cffi = "^23.1.0"
pyjwt = "^2.8.0"


[tool.poetry.group.dev.dependencies]
pytest = "^8.3.2"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
